Fler enhetstest av beräknaren (interim).

+++ src/Core.Domain/Business/IConsumptionCalculator.cs
@@ -5,6 +5,6 @@ namespace Branslekollen.Core.Domain.Business
 {
     public interface IConsumptionCalculator
     {
-        decimal? CalculateAverageConsumption(Vehicle vehicle, DateTime startDate, DateTime endDate);
+        double? CalculateAverageConsumption(Vehicle vehicle, DateTime startDate, DateTime endDate);
     }
 }
